requirements.txt
ultralytics
roboflow
yt-dlp
pyyaml
opencv-python
